requirements.txt
requests
pandas
python-dotenv
openai
streamlit
azure-functions
numpy
pytest
beads-mcp
